Finally! Boolean input keyboard works
It works because now the set / observable that stores the keys pressed is a global variable, so it doesn't rely on local variables resetting.
