+++ src/games/game-settings/player/Player.tsx
@@ -4,39 +4,17 @@ import { Spin, Tag } from 'antd'
 import getTypeSpecific from '../usePlayers/getTypeSpecific'
 import Observable from 'observables/lib/Observable'
 import PlayerIoType from '../PlayerIoType'
-import { useMemoOne } from 'use-memo-one'
 
 const Player = reactObserver<Props>((observe, { players, index, useGameSettingsOutput }) => {
   console.log('render player', index)
 
-  // const observable = observe(getTypeSpecific<Observable<boolean>>(
-  //   players,
-  //   index,
-  //   0,
-  //   useGameSettingsOutput,
-  //   PlayerIoType.BOOLEAN
-  // ))
-
-  const observableIsActive = useMemoOne(() => {
-    console.log('calculate')
-    return getTypeSpecific<Observable<boolean>>(
-      players,
-      index,
-      0,
-      useGameSettingsOutput,
-      PlayerIoType.BOOLEAN
-    )
-  }, [
+  const observable = observe(getTypeSpecific<Observable<boolean>>(
     players,
     index,
     0,
     useGameSettingsOutput,
     PlayerIoType.BOOLEAN
-  ])
-  // observableIsActive.addRemove.add(() => {})
-  const observable = observe(observableIsActive) // useMemoOne(() => observe(observableIsActive), [observableIsActive])
-  // observable?.addRemove.add(() => {})
-  console.log(observable)
+  ))
 
   return (
     <Spin spinning={observable === undefined}>
@@ -44,7 +22,7 @@ const Player = reactObserver<Props>((observe, { players, index, useGameSettingsO
         checked={observable !== undefined && observe(observable)}
         // checked
       >
-        Boolean Input Pressed
+        Player {index} Boolean Input Pressed
       </Tag.CheckableTag>
     </Spin>
   )
